+++ packages/pyapprox/src/pyapprox/util/sampling/__init__.py
@@ -1,9 +1,11 @@
 """Quasi-random sampling utilities."""
 
 from pyapprox.util.sampling.halton import HaltonSampler
+from pyapprox.util.sampling.latin_hypercube import LatinHypercubeSampler
 from pyapprox.util.sampling.sobol import SobolSampler
 
 __all__ = [
     "HaltonSampler",
+    "LatinHypercubeSampler",
     "SobolSampler",
 ]